telegram-bot/requirements.txt
python-telegram-bot
telegram
python_dotenv
Pillow
requests
